+++ src/api_entreprise/api.py
@@ -6,18 +6,13 @@ from api_entreprise.utils.url import join_fragments
 from . import logger
 from .models.config import Config
 from .models.donnees_etablissement import DonneesEtablissement
-from .models.numero_tva import NumeroTvaReponse
-from .models.chiffre_daffaires import ChiffreDaffairesResponse
-from .models.certification_rge import CertificationRgeResponse
+from .models.numero_tva import NumeroTvaHolder
+from .models.chiffre_daffaires import ChiffreDaffairesHolder
+from .models.certification_rge import CertificationRgeHolder
 from . import JSON_RESOURCE_IDENTIFIER
 from . import API_ENTREPRISE_VERSION
 
-from .handlers import (
-    _handle_httperr_429_ex,
-    _handle_bucketfull_ex,
-    _handle_httperr_404_returns_none,
-    _handle_response_in_error,
-)
+from .handlers import raw_call_handler
 
 from pyrate_limiter import BucketFullException
 
@@ -39,9 +34,9 @@ class ApiEntreprise:
             DonneesEtablissement | None: None si établissement non trouvé
         """
         json = self.raw_donnees_etablissement(siret)
-        return self._json_to_donnees_etab(json)
+        return self._json_to_dataclass(DonneesEtablissement, json)
 
-    def numero_tva_intercommunautaire(self, siren: str) -> NumeroTvaReponse | None:
+    def numero_tva_intercommunautaire(self, siren: str) -> NumeroTvaHolder | None:
         """Retourne le numéro de TVA intercommunautaire pour un siren donné
 
         Args:
@@ -51,9 +46,9 @@ class ApiEntreprise:
             NumeroTvaReponse|None: None si ressource non trouvée
         """
         json = self.raw_numero_tva_intercommunautaire(siren)
-        return self._json_to_numero_tva_response(json)
+        return self._json_to_dataclass(NumeroTvaHolder, json)
 
-    def chiffre_d_affaires(self, siret: str | int) -> list[ChiffreDaffairesResponse]:
+    def chiffre_d_affaires(self, siret: str | int) -> list[ChiffreDaffairesHolder]:
         """Retourne le chiffre d'affaires des trois derniers exercices
         faites auprès de la DGFIP
 
@@ -62,58 +57,66 @@ class ApiEntreprise:
         """
 
         json = self.raw_chiffre_d_affaires(siret)
-        return self._json_to_chiffre_d_affaires(json)
+        return self._json_to_dataclass(ChiffreDaffairesHolder, json, many=True)
 
-    def certifications_rge(self, siret: str | int) -> list[CertificationRgeResponse]:
+    def certifications_rge(self, siret: str | int) -> list[CertificationRgeHolder]:
         """Retourne les certifications RGE (Reconnu Garant de l'Environnement) d'un établissement
 
         Returns:
             list[CertificationRgeResponse]:
         """
         json = self.raw_certification_rge(siret)
-        return self._json_to_certification_rge(json)
+        return self._json_to_dataclass(CertificationRgeHolder, json, many=True)
 
-    @_handle_response_in_error
-    @_handle_httperr_404_returns_none
-    @_handle_httperr_429_ex
-    @_handle_bucketfull_ex
     def raw_donnees_etablissement(self, siret: str) -> dict | None:
-        response = self._donnees_etablissement(siret)
-        response.raise_for_status()
-        json = response.json()
-        return json
+        f = lambda: self._donnees_etablissement(siret)
+        return self._raw(f)
 
-    @_handle_response_in_error
-    @_handle_httperr_404_returns_none
-    @_handle_httperr_429_ex
-    @_handle_bucketfull_ex
     def raw_numero_tva_intercommunautaire(self, siren: str | int) -> dict | None:
-        response = self._numero_tva_intercommunautaire(siren)
-        response.raise_for_status()
-        json = response.json()
-        return json
+        f = lambda: self._numero_tva_intercommunautaire(siren)
+        return self._raw(f)
 
-    @_handle_response_in_error
-    @_handle_httperr_404_returns_none
-    @_handle_httperr_429_ex
-    @_handle_bucketfull_ex
     def raw_chiffre_d_affaires(self, siret: str | int) -> dict | None:
-        response = self._chiffre_d_affaires(siret)
-        response.raise_for_status()
-        json = response.json()
-        return json
+        f = lambda: self._chiffre_d_affaires(siret)
+        return self._raw(f)
 
-    @_handle_response_in_error
-    @_handle_httperr_404_returns_none
-    @_handle_httperr_429_ex
-    @_handle_bucketfull_ex
     def raw_certification_rge(self, siret: str | int) -> dict | None:
-        response = self._certification_rge(siret)
+        f = lambda: self._certification_rge(siret)
+        return self._raw(f)
+
+    @raw_call_handler
+    def _raw(self, f) -> dict | None:
+        response: requests.Response = f()
         response.raise_for_status()
         json = response.json()
         return json
 
     def _donnees_etablissement(self, siret) -> requests.Response:
+        url = join_fragments(self._base_url, f"insee/sirene/etablissements/{siret}")
+        return self._perform_get(url)
+
+    def _numero_tva_intercommunautaire(self, siren: str | int) -> requests.Response:
+        url = join_fragments(
+            self._base_url,
+            f"european_commission/unites_legales/{siren}/numero_tva",
+        )
+        return self._perform_get(url)
+
+    def _chiffre_d_affaires(self, siret: str | int) -> requests.Response:
+        url = join_fragments(
+            self._base_url,
+            f"dgfip/etablissements/{siret}/chiffres_affaires",
+        )
+        return self._perform_get(url)
+
+    def _certification_rge(self, siret: str | int) -> requests.Response:
+        url = join_fragments(
+            self._base_url,
+            f"ademe/etablissements/{siret}/certification_rge",
+        )
+        return self._perform_get(url)
+
+    def _perform_get(self, url) -> requests.Response:
         # On utilise un lock avec le ratelimiter
         # car ce dernier se comporte mal en situation
         # d'un grand nombre de tâches en //
@@ -122,60 +125,7 @@ class ApiEntreprise:
             self._ratelimiter.ratelimit(JSON_RESOURCE_IDENTIFIER) as _,
         ):
             response = requests.get(
-                join_fragments(self._base_url, f"insee/sirene/etablissements/{siret}"),
-                headers=self._auth_headers,
-                params=self._query_params,
-            )
-            self._empty_ratelimiter_if_429(response)
-
-            return response
-
-    def _numero_tva_intercommunautaire(self, siren: str | int) -> requests.Response:
-        with (
-            _ratelimiterlock as _,
-            self._ratelimiter.ratelimit(JSON_RESOURCE_IDENTIFIER) as _,
-        ):
-            # https://entreprise.api.gouv.fr/v3/european_commission/unites_legales/{siren}/numero_tva
-            response = requests.get(
-                join_fragments(
-                    self._base_url,
-                    f"european_commission/unites_legales/{siren}/numero_tva",
-                ),
-                headers=self._auth_headers,
-                params=self._query_params,
-            )
-            self._empty_ratelimiter_if_429(response)
-
-            return response
-
-    def _chiffre_d_affaires(self, siret: str | int) -> requests.Response:
-        with (
-            _ratelimiterlock as _,
-            self._ratelimiter.ratelimit(JSON_RESOURCE_IDENTIFIER) as _,
-        ):
-            # https://entreprise.api.gouv.fr/v3/dgfip/etablissements/{siret}/chiffres_affaires
-            response = requests.get(
-                join_fragments(
-                    self._base_url, f"dgfip/etablissements/{siret}/chiffres_affaires"
-                ),
-                headers=self._auth_headers,
-                params=self._query_params,
-            )
-            self._empty_ratelimiter_if_429(response)
-
-            return response
-
-    def _certification_rge(self, siret: str | int) -> requests.Response:
-        with (
-            _ratelimiterlock as _,
-            self._ratelimiter.ratelimit(JSON_RESOURCE_IDENTIFIER) as _,
-        ):
-            # https://entreprise.api.gouv.fr/v3/ademe/etablissements/{siret}/certification_rge
-
-            response = requests.get(
-                join_fragments(
-                    self._base_url, f"ademe/etablissements/{siret}/certification_rge"
-                ),
+                url,
                 headers=self._auth_headers,
                 params=self._query_params,
             )
@@ -199,29 +149,14 @@ class ApiEntreprise:
     def _base_url(self):
         return join_fragments(self._config.base_url, API_ENTREPRISE_VERSION)
 
-    def _json_to_donnees_etab(self, json):
-        schema = DonneesEtablissement.ma_schema
+    def _json_to_dataclass(self, cls, json, many=False):
+        if many:
+            schema = cls.ma_schema_many
+        else:
+            schema = cls.ma_schema
 
-        donnees = schema.load(json["data"])
-        return donnees
-
-    def _json_to_numero_tva_response(self, json):
-        schema = NumeroTvaReponse.ma_schema
-
-        tva = schema.load(json["data"])
-        return tva
-
-    def _json_to_chiffre_d_affaires(self, json):
-        schema = ChiffreDaffairesResponse.ma_schema_many
-
-        ca = schema.load(json["data"])
-        return ca
-
-    def _json_to_certification_rge(self, json):
-        schema = CertificationRgeResponse.ma_schema_many
-
-        certif = schema.load(json["data"])
-        return certif
+        dc = schema.load(json["data"])
+        return dc
 
     def _empty_ratelimiter_if_429(self, response: requests.Response):
         if response.status_code == 429: