+++ tests/test_numero_tva.py
@@ -3,7 +3,7 @@ from .fixtures import *
 
 from .fixtures import vcr_folder
 
-from api_entreprise import ApiEntreprise
+from api_entreprise import ApiEntreprise, NumeroTvaHolder
 
 
 @vcr.use_cassette(f"{vcr_folder}/tva.megalis.yml")
@@ -16,3 +16,4 @@ def test_raw_tva(api: ApiEntreprise):
 def test_tva(api: ApiEntreprise):
     tva = api.numero_tva_intercommunautaire(821936291)
     assert tva is not None
+    assert isinstance(tva, NumeroTvaHolder)